+++ packages/sdk/src/ui/ui.js
@@ -47,6 +47,7 @@ import useLoadable from './use_loadable';
 import {useRecordIds, useRecords, useRecordById} from './use_records';
 import useBase from './use_base';
 import useSession from './use_session';
+import useSettingsButton from './use_settings_button';
 import useWatchable from './use_watchable';
 import useViewport from './use_viewport';
 import useGlobalConfig from './use_global_config';
@@ -102,6 +103,7 @@ const UI = {
     withHooks,
     useBase,
     useSession,
+    useSettingsButton,
     useViewport,
     useWatchable,
     useLoadable,